compass revamp
Revamps the compass code to better adapt for its compass, compass code is accurate to basically near verticle tilts.

Automatically does a running max-min, but will have to use floats to do the calculation of the heading (as the mag sensor's x,y, and z do not have the same magnitude as each other).

The comp.ino code will return the system's current heading, given it's current knowledge of the max and mins of the sensor (as this changes from sensor to sensor and sometimes day to day, it would be better to use this method).